Make history's height semi-flexible
When setting the application's height to `auto` (in `DEFAULT_OPTIONS`), then setting the history's height to 100% results in a minimal height of the history, as there is no fixed reference value to base the 100% on.

This change now sets the history's minimum height to the equivalent of players being displayed.
The height values are fixed for 1-6 players (not counting yourself).
Having 7 players or more results in a scrollbar showing.

Having only a single user (excluding user role `none`) in the world (AKA "only yourself") is currently not handled, since talking to yourself via the messenger is not supported.

This solution is clunky but acceptable as the goal is to move to a vertical tab-based view anyway.

This closes #5