✨v13 support: Chatbar btn handles new core styling
Foundry VTT v13 introduces significant changes to its UI, preventing
adding the buttons to open its window to the sidebar chat tab and the
scene controls toolbar.

This change adjusts the former accordingly, and introduces moving the
button to the `chat-notification` if the sidebar is closed, similar to
core's roll privacy buttons.
As the button is now always visible even if the sidebar chat tab is not
visible, with v13 this becomes the single UI element that's added and
non-optional.
Adding buttons to the scene controls toolbar which don't relate to the
canvas is highly discouraged in the community anyway.

This closes #62
This closes #89